Shaders in renderer and Camera2D

+++ hypermath/include/trigonometry/trig_functions.h
@@ -8,7 +8,7 @@
 
 namespace hpm {
 	inline constexpr float degToRad(float degrees) {
-		return degrees * (M_PI / 180.0f);
+		return degrees * (static_cast<float>(M_PI) / 180.0f);
 	}
 }
 #endif
