+++ parallax/include/textures/TextureAtlas.h
@@ -34,9 +34,9 @@ namespace prx {
 		// Rebuilds actual rendering API texture with new added regions.
 		void update();
 
-		inline unsigned int getUsage()  const { return m_TextureAtlas->used;   };
-		inline unsigned int getWidth()  const { return m_TextureAtlas->width;  };
-		inline unsigned int getHeight() const { return m_TextureAtlas->height; };
+		inline unsigned int getUsage()  const { return static_cast<unsigned int>(m_TextureAtlas->used);   };
+		inline unsigned int getWidth()  const { return static_cast<unsigned int>(m_TextureAtlas->width);  };
+		inline unsigned int getHeight() const { return static_cast<unsigned int>(m_TextureAtlas->height); };
 	};
 
 }
